prueba

+++ app/Http/Controllers/LoginController.php
@@ -7,8 +7,16 @@ use Illuminate\Support\Facades\Auth;
 
 class LoginController extends Controller
 {
-    public function login(Token $token)
+    public function login($token)
     {
+        $token = Token::where('token', $token)->first();
+
+        if ($token == null) {
+            alert('Este enlace ya expiró, por favor solicite otro', 'danger');
+
+            return redirect()->route('token');
+        }
+
         Auth::login($token->user);
 
         $token->delete();
